refactor(frontend): refactor vector view layers
- move view layer definitions from `src/state` to `src/config`.
- use `viewLayer.renderTooltip` to show tooltips for interactive layers.
- remove the redundant `Component` property from vector layers.
