+++ frontend/src/config/regions/boundaries-view-layer.ts
@@ -1,6 +1,11 @@
+import { createElement } from 'react';
+
+import { VectorTarget } from 'lib/data-map/types';
+import { ViewLayer } from 'lib/data-map/view-layers';
+
 import { regionBoundariesDeckLayer } from './region-boundaries-deck-layer';
 import { RegionLevel } from './metadata';
-import { ViewLayer } from 'lib/data-map/view-layers';
+import { RegionHoverDescription } from './RegionHoverDescription';
 
 export function regionBoundariesViewLayer(regionLevel: RegionLevel): ViewLayer {
   return {
@@ -12,5 +17,8 @@ export function regionBoundariesViewLayer(regionLevel: RegionLevel): ViewLayer {
       regionLevel,
     },
     fn: () => regionBoundariesDeckLayer(regionLevel),
+    renderTooltip({ target }: { target: VectorTarget }) {
+      return createElement(RegionHoverDescription, { key: this.id, target, viewLayer: this });
+    },
   };
 }
