+++ frontend/src/state/layers/modules/terrestrial.ts
@@ -1,25 +1,18 @@
-import { DataFilterExtension } from '@deck.gl/extensions';
 import { TERRESTRIAL_LANDUSE_COLORS } from 'config/solutions/colors';
+import { terrestrialViewLayer } from 'config/solutions/terrestrial-view-layer';
 import { ViewLayer, FieldSpec, ColorSpec } from 'lib/data-map/view-layers';
 import { selector } from 'recoil';
 import { sectionStyleValueState, sectionVisibilityState } from 'state/sections';
 import {
-  TerrestrialFilters,
   terrestrialFiltersState,
   TerrestrialLocationFilters,
 } from 'state/solutions/terrestrial-filters';
 import { colorMap } from 'lib/color-map';
 import { VECTOR_COLOR_MAPS } from 'config/color-maps';
-import { VectorTarget } from 'lib/data-map/types';
 import { featureProperty } from 'lib/deck/props/data-source';
-import { dataColorMap } from 'lib/deck/props/color-map';
-import { border, fillColor } from 'lib/deck/props/style';
 import { Accessor } from 'lib/deck/props/getters';
 import { LandUseOption, TerrestrialLocationFilterType } from 'config/solutions/domains';
 import { truthyKeys } from 'lib/helpers';
-import { selectableMvtLayer } from 'lib/deck/layers/selectable-mvt-layer';
-import { getTerrestrialDataFormats } from 'config/solutions/data-formats';
-import { getSolutionsDataAccessor } from 'config/solutions/data-access';
 import { landuseFilterState } from 'state/solutions/landuse-tree';
 
 export function landuseColorMap(x: string) {
@@ -81,15 +74,6 @@ export const terrestrialFieldSpecState = selector<FieldSpec>({
   },
 });
 
-function landuseFilterValue(p, landuseFilters: Set<LandUseOption>) {
-  return landuseFilters.has(p.landuse_desc) ? 1 : 0;
-}
-
-function locationFilterValue(p, locationFiltersKeys: TerrestrialLocationFilterType[]) {
-  //eslint-disable-next-line eqeqeq -- values are currently sometimes 1 and sometimes true
-  return locationFiltersKeys.every((key) => p[key] == true) ? 1 : 0;
-}
-
 const landuseFilterSetState = selector<Set<LandUseOption>>({
   key: 'landuseFilterSetState',
   get: ({ get }) => new Set(truthyKeys(get(landuseFilterState))),
@@ -104,34 +88,6 @@ const locationFilterKeysState = selector<TerrestrialLocationFilterType[]>({
   key: 'locationFilterKeysState',
   get: ({ get }) => truthyKeys(get(locationFilterState)),
 });
-
-function terrestrialFilters(
-  filters: TerrestrialFilters,
-  landuseFilterSet: Set<LandUseOption>,
-  locationFilterKeys: TerrestrialLocationFilterType[],
-  doFilter: boolean = true,
-) {
-  return {
-    getFilterValue: ({ properties }) => [
-      doFilter ? landuseFilterValue(properties, landuseFilterSet) : 1,
-      doFilter ? properties.slope_degrees : 1,
-      doFilter ? properties.elevation_m : 1,
-      doFilter ? locationFilterValue(properties, locationFilterKeys) : 1,
-    ],
-    filterRange: [
-      [1, 1],
-      doFilter ? [...filters.slope_degrees] : [1, 1],
-      doFilter ? [...filters.elevation_m] : [1, 1],
-      [1, 1],
-    ],
-
-    updateTriggers: {
-      getFilterValue: [doFilter, landuseFilterSet, locationFilterKeys],
-    },
-
-    extensions: [new DataFilterExtension({ filterSize: 4 })],
-  };
-}
 
 export const terrestrialLayerState = selector<ViewLayer>({
   key: 'terrestrialLayerState',
@@ -157,85 +113,14 @@ export const terrestrialLayerState = selector<ViewLayer>({
     const landuseFilterSet = get(landuseFilterSetState);
     const locationFilterKeys = get(locationFilterKeysState);
 
-    return {
-      id: 'terrestrial',
-      group: null,
-      interactionGroup: 'solutions',
-      styleParams: colorSpec && {
-        colorMap: {
-          fieldSpec,
-          colorSpec,
-        },
-      },
-      dataAccessFn: getSolutionsDataAccessor,
-      dataFormatsFn: getTerrestrialDataFormats,
-      fn: ({ deckProps, zoom, selection }) => {
-        const switchoverZoom = 14.5;
-        const target = selection?.target as VectorTarget;
-        const selectedFeatureId = target?.feature.id;
-
-        return [
-          selectableMvtLayer(
-            {
-              selectionOptions: {
-                selectedFeatureId,
-                polygonOffset: -1000,
-              },
-            },
-            deckProps,
-            {
-              id: `${deckProps.id}@points`,
-              data: '/vector/data/natural_terrestrial_combined_points.json',
-              visible: zoom < switchoverZoom,
-              binary: false,
-              filled: true,
-              pointAntialiasing: false,
-              stroked: false,
-            },
-            {
-              getPointRadius: 35,
-              pointRadiusUnit: 'meters',
-              pointRadiusMinPixels: 4,
-              pointRadiusMaxPixels: 7,
-            },
-            fillColor(dataColorMap(dataFn, colorFn)),
-            terrestrialFilters(
-              filters,
-              landuseFilterSet,
-              locationFilterKeys,
-              zoom < switchoverZoom,
-            ),
-          ),
-          selectableMvtLayer(
-            {
-              selectionOptions: {
-                selectedFeatureId,
-                polygonOffset: -1000,
-              },
-            },
-            deckProps,
-            {
-              data: '/vector/data/natural_terrestrial_combined.json',
-              minZoom: 14,
-              visible: zoom >= switchoverZoom,
-              binary: false,
-              filled: true,
-
-              getLineWidth: 1,
-              lineWidthUnit: 'pixels',
-              lineWidthMinPixels: 1,
-            },
-            border(),
-            fillColor(dataColorMap(dataFn, colorFn)),
-            terrestrialFilters(
-              filters,
-              landuseFilterSet,
-              locationFilterKeys,
-              zoom >= switchoverZoom,
-            ),
-          ),
-        ];
-      },
-    };
+    return terrestrialViewLayer({
+      fieldSpec,
+      colorSpec,
+      dataFn,
+      colorFn,
+      filters,
+      landuseFilterSet,
+      locationFilterKeys,
+    });
   },
 });
